+++ src/components/EntireProducts.js
@@ -1,5 +1,5 @@
 import "./EntireProducts.css";
-import heart from "../assets/Icon.svg";
+import heart from "../assets/icons/Icon.svg";
 
 function ProductsList({ item }) {
   const { images, name, price, favoriteCount } = item;
